requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
mcp>=0.1.0
streamlit>=1.50.0
pydantic-evals>=0.1.0

